refactor(engine): merge QueryBinding into TableBinding
Remove QueryBinding interface and integrate its query methods (get,
count, scan) into TableBinding. This eliminates a separate abstraction
for queries and unifies mutation and query operations under a single
table-scoped binding.

- Add get/count/scan to TableBinding, ScanFilter to binding package
- Implement query methods in V2BackedTableBinding (absorb V2BackedQueryBinding logic)
- Add query stubs to NilTableBinding
- Change ActionbaseQueryExecutor to accept (database, table) -> TableBinding resolver
- Delete QueryBinding, QueryScanFilter, V2BackedQueryBinding
