+++ engine/src/main/kotlin/com/kakao/actionbase/engine/binding/ScanFilter.kt
@@ -1,9 +1,9 @@
-package com.kakao.actionbase.engine.query
+package com.kakao.actionbase.engine.binding
 
 import com.kakao.actionbase.v2.core.metadata.Direction
 import com.kakao.actionbase.v2.engine.sql.WherePredicate
 
-data class QueryScanFilter(
+data class ScanFilter(
     val sourceSet: Set<Any>,
     val direction: Direction,
     val limit: Int,
