+++ engine/src/main/kotlin/com/kakao/actionbase/v2/engine/v3/NilTableBinding.kt
@@ -4,8 +4,13 @@ import com.kakao.actionbase.core.edge.MutationKey
 import com.kakao.actionbase.core.metadata.common.ModelSchema
 import com.kakao.actionbase.core.state.State
 import com.kakao.actionbase.engine.binding.MutationRecordsSummary
+import com.kakao.actionbase.engine.binding.ScanFilter
 import com.kakao.actionbase.engine.binding.TableBinding
 import com.kakao.actionbase.engine.metadata.MutationMode
+import com.kakao.actionbase.v2.core.metadata.Direction
+import com.kakao.actionbase.v2.core.types.StructType
+import com.kakao.actionbase.v2.engine.sql.DataFrame
+import com.kakao.actionbase.v2.engine.sql.StatKey
 
 import reactor.core.publisher.Mono
 
@@ -31,7 +36,24 @@ class NilTableBinding(
 
     override fun handleMutationError(error: Throwable) {}
 
+    override fun get(
+        source: List<Any>,
+        target: List<Any>,
+        stats: Set<StatKey>,
+    ): Mono<DataFrame> = EMPTY_DATAFRAME
+
+    override fun count(
+        source: Set<Any>,
+        direction: Direction,
+    ): Mono<DataFrame> = EMPTY_DATAFRAME
+
+    override fun scan(
+        filter: ScanFilter,
+        stats: Set<StatKey>,
+    ): Mono<DataFrame> = EMPTY_DATAFRAME
+
     private companion object {
+        val EMPTY_DATAFRAME: Mono<DataFrame> = Mono.just(DataFrame(emptyList(), StructType(emptyArray())))
         const val IDLE = "IDLE"
     }
 }
